01-streaming-realtime/kafka-to-delta-lake/requirements.txt
# Core Streaming Libraries
pyspark==3.5.0
delta-spark==3.0.0

# Kafka Integration
kafka-python==2.0.2

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Configuration Management
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring & Metrics
prometheus-client==0.19.0

# Logging
loguru==0.7.2

# Utilities
click==8.1.7
